1. This commit moves the (auto)follow functionality from the notification app to the follows app. 2. Introduces a new SubmitJourneyNotification that notifies the author of the journey that its submission succeeded. 3. It moves the notification sending on status changes from the pre_save to the post_save method of ideas.
