+++ apps/notifications/templates/advocate_europe_notifications/emails/submit_journey_notification.en.email
@@ -1,0 +1,20 @@
+{% extends 'email_base.'|add:part_type %}
+
+{% block subject %}
+You have sucessfully submitted your journey "{{ journey.title }}"
+{% endblock %}
+
+{% block headline %}
+Sucessfully submission!
+{% endblock  %}
+
+{% block content %}
+You have submitted your journey "{{ journey.title }}" to {{ journey.idea.type }} "{{ journey.idea.idea_title }}".
+{% endblock %}
+
+{% block cta_url %}{{ email.get_host }}{{ journey.get_absolute_url }}{% endblock %}
+{% block cta_label %}Visit your journey{% endblock %}
+
+{% block reason %}
+This email was sent to {{ receiver }}. You receive this email because you have submitted a journey using this email address. If you have any further questions, please contact us via info@advocate-europe.eu.
+{% endblock %}
